Feature/task with recycler view

+++ app/src/main/java/com/example/lomakincountriesapp/data/Country.kt
@@ -5,6 +5,6 @@ data class Country(
     var capital: ArrayList<String>,
     var languages: Map<String?, String?>,
     var area: Double,
-    var population: Int,
+    var population: Long,
     var flags: Flags?,
 )
